requirements.txt
flask
apscheduler
python-telegram-bot
requests
sentence-transformers
openai
